Добавила функцию renderReviewElement в single-review модуль, подправила метод onReviewQuizClick
